requirements.txt
line-bot-sdk
flask
python-dateutil
requests
url-normalize 
beautifulsoup4
bs4
Beautifulsoup4